# specs/1-todo-fullstack-auth/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Application API
  description: API for Todo Full-Stack Web Application with Authentication
  version: 1.0.0
  contact:
    name: Todo App Development Team

servers:
  - url: https://api.todoapp.com/v1
    description: Production server
  - url: https://staging-api.todoapp.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
    description: Development server

paths:
  /auth/register:
    post:
      summary: Register a new user
      description: Creates a new user account with email and password
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                  example: user@example.com
                password:
                  type: string
                  format: password
                  description: User's password
                  example: securePassword123
      responses:
        '201':
          description: User successfully registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User registered successfully
                  user_id:
                    type: integer
                    example: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  /auth/login:
    post:
      summary: Login user
      description: Authenticates user and returns JWT token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                  example: user@example.com
                password:
                  type: string
                  format: password
                  description: User's password
                  example: securePassword123
      responses:
        '200':
          description: User successfully authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT access token
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  token_type:
                    type: string
                    description: Type of token
                    example: bearer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /auth/logout:
    post:
      summary: Logout user
      description: Invalidates the current user session
      tags:
        - Authentication
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User successfully logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Successfully logged out

  /tasks:
    get:
      summary: Get user's tasks
      description: Retrieve all tasks belonging to the authenticated user
      tags:
        - Tasks
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of user's tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: Create a new task
      description: Create a new task for the authenticated user
      tags:
        - Tasks
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                  description: Title of the task
                  example: Buy groceries
                description:
                  type: string
                  description: Detailed description of the task
                  example: Buy milk, bread, and eggs
                completed:
                  type: boolean
                  description: Whether the task is completed
                  default: false
                due_date:
                  type: string
                  format: date-time
                  description: Due date for the task
                  example: '2023-12-31T23:59:59Z'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /tasks/{task_id}:
    get:
      summary: Get a specific task
      description: Retrieve a specific task belonging to the authenticated user
      tags:
        - Tasks
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: ID of the task to retrieve
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a task
      description: Update a specific task belonging to the authenticated user
      tags:
        - Tasks
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: ID of the task to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: Updated title of the task
                  example: Buy groceries
                description:
                  type: string
                  description: Updated description of the task
                  example: Buy milk, bread, eggs, and fruit
                completed:
                  type: boolean
                  description: Updated completion status
                  example: true
                due_date:
                  type: string
                  format: date-time
                  description: Updated due date for the task
                  example: '2023-12-31T23:59:59Z'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      summary: Delete a task
      description: Delete a specific task belonging to the authenticated user
      tags:
        - Tasks
      security:
        - BearerAuth: []
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
          description: ID of the task to delete
      responses:
        '204':
          description: Task deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the task
          example: 1
        user_id:
          type: integer
          description: ID of the user who owns this task
          example: 1
        title:
          type: string
          description: Title of the task
          example: Buy groceries
        description:
          type: string
          description: Detailed description of the task
          example: Buy milk, bread, and eggs
        completed:
          type: boolean
          description: Whether the task is completed
          example: false
        created_at:
          type: string
          format: date-time
          description: When the task was created
          example: '2023-01-01T12:00:00Z'
        updated_at:
          type: string
          format: date-time
          description: When the task was last updated
          example: '2023-01-01T12:00:00Z'
        due_date:
          type: string
          format: date-time
          description: Due date for the task
          example: '2023-12-31T23:59:59Z'
      required:
        - id
        - user_id
        - title
        - completed
        - created_at
        - updated_at

  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: Invalid input data
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: Not authenticated
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: Resource not found
    Conflict:
      description: Conflict - such as email already registered
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: Email already registered